Move TrashcanOpen into seperate file
